Fixes #2568 by delivering messages before probing conversations

+++ msi.gaml.extensions.fipa/src/msi/gaml/extensions/fipa/ConversationType.java
@@ -1,12 +1,12 @@
 /*********************************************************************************************
- * 
- * 
+ *
+ *
  * 'ConversationType.java', in plugin 'msi.gaml.extensions.fipa', is part of the source code of the GAMA modeling and
  * simulation platform. (c) 2007-2014 UMI 209 UMMISCO IRD/UPMC & Partners
- * 
+ *
  * Visit https://code.google.com/p/gama-platform/ for license information and developers contact.
- * 
- * 
+ *
+ *
  **********************************************************************************************/
 package msi.gaml.extensions.fipa;
 
@@ -33,7 +33,7 @@ import msi.gaml.types.Types;
 public class ConversationType extends GamaContainerType<Conversation> {
 
 	public final static String CONVERSATION_STR = "conversation";
-	public final static short CONV_ID = 98;
+	public final static int CONV_ID = 98;
 
 	public ConversationType() {}
